Added code to only owner can manipulate his projects

+++ homepageBackendSolution/homepageBackend/Controllers/V1/ProjectsController.cs
@@ -5,10 +5,12 @@ using homepageBackend.Contracts.V1;
 using homepageBackend.Contracts.V1.Requests;
 using homepageBackend.Contracts.V1.Responses;
 using homepageBackend.Domain;
+using homepageBackend.Extensions;
 using homepageBackend.Services;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Npgsql.EntityFrameworkCore.PostgreSQL.Query.Expressions.Internal;
 
 namespace homepageBackend.Controllers
 {
@@ -45,12 +47,15 @@ namespace homepageBackend.Controllers
         [Route(ApiRoutes.Projects.Update)]
         public async Task<IActionResult> Update([FromRoute] Guid projectId, [FromBody] UpdateProjectRequest request)
         {
-            HttpContext.User.Claims.Single(x => x.Type == "id");
-            var project = new Project
+            var userOwnsProject = await _projectService.UserOwnsPostAsync(projectId, HttpContext.GetUserId());
+
+            if (!userOwnsProject)
             {
-                Id = projectId,
-                name = request.Name
-            };
+                return BadRequest(new {error = "You do not own this post"});
+            }
+
+            var project = await _projectService.GetProjectIdAsync(projectId);
+            project.Name = request.Name;
 
             var updated = await _projectService.UpdateProjectAsync(project);
 
@@ -63,9 +68,15 @@ namespace homepageBackend.Controllers
         [Route(ApiRoutes.Projects.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid projectId)
         {
+            var userOwnsProject = await _projectService.UserOwnsPostAsync(projectId, HttpContext.GetUserId());
+
+            if (!userOwnsProject)
+            {
+                return BadRequest(new {error = "You do not own this post"});
+            }
+
             var deleted = await _projectService.DeleteProjectAsync(projectId);
 
-            
             if (deleted)
                 return NoContent(); // 204
 
@@ -76,8 +87,12 @@ namespace homepageBackend.Controllers
         [Route(ApiRoutes.Projects.Create)]
         public async Task<IActionResult> Create([FromBody] CreateProjectRequest projectRequest)
         {
-            var project = new Project {name = projectRequest.Name};
-            
+            var project = new Project
+            {
+                Name = projectRequest.Name,
+                UserId = HttpContext.GetUserId()
+            };
+
             await _projectService.CreateProjectAsync(project);
 
             var baseUrl = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host.ToUriComponent()}";
@@ -85,7 +100,7 @@ namespace homepageBackend.Controllers
 
 
             var response = new ProjectResponse {Id = project.Id};
-            return Created(locationUri, project);
+            return Created(locationUri, response);
         }
     }
 }
